functions/ingestion/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-secret-manager==2.*
langchain==0.2.*
langchain-community==0.2.*
chromadb==0.5.*
sentence-transformers==3.*
PyGithub==2.*
requests==2.*
google-genai